Extract JarSignatureCollector into its own file
